+++ p5/sketch/Vispy2DRenderer/shaders/common/default.frag
@@ -1,0 +1,6 @@
+varying vec4 frag_color;
+
+void main()
+{
+    gl_FragColor = frag_color;
+}
