+++ p5/sketch/Vispy3DRenderer/shaders/3d/normal.frag
@@ -1,0 +1,7 @@
+precision mediump float;
+varying vec3 v_normal;
+
+void main()
+{
+    gl_FragColor = vec4(normalize(v_normal), 1.0);
+}
